fix: 전단 미리보기 상품 할인가 타입을 BigDecimal로 통일

+++ api/owner-api/src/main/kotlin/kr/dongchimi/api/owner/flyer/response/FlyerPreviewProductResponse.kt
@@ -2,6 +2,7 @@ package kr.dongchimi.api.owner.flyer.response
 
 import io.swagger.v3.oas.annotations.media.Schema
 import kr.dongchimi.core.product.Product
+import java.math.BigDecimal
 
 data class FlyerPreviewProductResponse(
     @Schema(description = "상품 id")
@@ -11,7 +12,7 @@ data class FlyerPreviewProductResponse(
     @Schema(description = "상품 썸네일 이미지 URL")
     val thumbnailUrl: String,
     @Schema(description = "할인가")
-    val discountedPrice: Int,
+    val discountedPrice: BigDecimal,
     @Schema(description = "할인율(%). 0이면 할인 없음")
     val discountRate: Int,
 ) {
@@ -19,7 +20,7 @@ data class FlyerPreviewProductResponse(
         productId = product.id,
         name = product.name,
         thumbnailUrl = product.thumbnailUrl,
-        discountedPrice = product.price.discountedPrice.toInt(),
+        discountedPrice = product.price.discountedPrice,
         discountRate = product.price.discountRate(),
     )
 }
